src: generalize sorted coords to counts
