+++ legate_sparse_cpp.cmake
@@ -108,6 +108,7 @@ list(APPEND legate_sparse_SOURCES
   src/sparse/array/conv/csr_to_dense.cc
   src/sparse/array/conv/dense_to_csr.cc
   src/sparse/array/conv/pos_to_coordinates.cc
+  src/sparse/array/conv/sorted_coords_to_counts.cc
   src/sparse/array/csc/spmv.cc
   src/sparse/array/csr/add.cc
   src/sparse/array/csr/get_diagonal.cc
@@ -133,6 +134,7 @@ if(Legion_USE_OpenMP)
     src/sparse/array/conv/csr_to_dense_omp.cc
     src/sparse/array/conv/dense_to_csr_omp.cc
     src/sparse/array/conv/pos_to_coordinates_omp.cc
+    src/sparse/array/conv/sorted_coords_to_counts_omp.cc
     src/sparse/array/csc/spmv_omp.cc
     src/sparse/array/csr/add_omp.cc
     src/sparse/array/csr/get_diagonal_omp.cc
@@ -155,6 +157,7 @@ if(Legion_USE_CUDA)
     src/sparse/array/conv/csr_to_dense.cu
     src/sparse/array/conv/dense_to_csr.cu
     src/sparse/array/conv/pos_to_coordinates.cu
+    src/sparse/array/conv/sorted_coords_to_counts.cu
     src/sparse/array/csc/spmv.cu
     src/sparse/array/csr/get_diagonal.cu
     src/sparse/array/csr/add.cu
