+++ src/sparse/array/conv/sorted_coords_to_counts.h
@@ -1,0 +1,43 @@
+/* Copyright 2022 NVIDIA Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+#pragma once
+
+#include "sparse.h"
+#include "sparse_c.h"
+#include "legate.h"
+
+namespace sparse {
+
+struct SortedCoordsToCountsArgs {
+  const legate::Store& output;
+  const legate::Store& input;
+  const int64_t max_vals;
+};
+
+class SortedCoordsToCounts : public SparseTask<SortedCoordsToCounts> {
+ public:
+  static const int TASK_ID = LEGATE_SPARSE_SORTED_COORDS_TO_COUNTS;
+  static void cpu_variant(legate::TaskContext& ctx);
+#ifdef LEGATE_USE_OPENMP
+  static void omp_variant(legate::TaskContext& ctx);
+#endif
+#ifdef LEGATE_USE_CUDA
+  static void gpu_variant(legate::TaskContext& ctx);
+#endif
+};
+
+}  // namespace sparse
