+++ src/tasks.cu
@@ -1070,47 +1070,6 @@ void CSCSDDMM::gpu_variant(legate::TaskContext& ctx)
   CHECK_CUDA_STREAM(stream);
 }
 
-template <typename T>
-__global__ void scatter_reduce(size_t elems, T out, coord_t* keys, uint64_t* counts)
-{
-  auto idx = global_tid_1d();
-  if (idx >= elems) return;
-  out[keys[idx]] <<= counts[idx];
-}
-
-void SortedCoordsToCounts::gpu_variant(legate::TaskContext& ctx)
-{
-  auto& output  = ctx.reductions()[0];
-  auto& input   = ctx.inputs()[0];
-  auto max_vals = ctx.scalars()[0].value<int64_t>();
-  auto dom      = input.domain();
-  // Return if there are no coordinates to process.
-  if (output.domain().empty() || dom.empty()) return;
-  auto in  = input.read_accessor<coord_ty, 1>();
-  auto out = output.reduce_accessor<SumReduction<uint64_t>, true /* exclusive */, 1>();
-
-  // TODO (rohany): We could make a call to unique before this to avoid allocating
-  //  O(rows/cols) space. For now this shouldn't be too much of a problem. Unfortunately
-  //  unique counting has just recently landed, and I'm not sure when our thrust version
-  //  will pick up the change: https://github.com/NVIDIA/thrust/issues/1612.
-  auto kind   = Memory::GPU_FB_MEM;
-  auto stream = get_cached_stream();
-  ThrustAllocator alloc(kind);
-  auto policy = thrust::cuda::par(alloc).on(stream);
-  DeferredBuffer<coord_t, 1> keys({0, max_vals - 1}, kind);
-  DeferredBuffer<uint64_t, 1> counts({0, max_vals - 1}, kind);
-  auto result = thrust::reduce_by_key(policy,
-                                      in.ptr(dom.lo()),
-                                      in.ptr(dom.lo()) + dom.get_volume(),
-                                      thrust::make_constant_iterator(1),
-                                      keys.ptr(0),
-                                      counts.ptr(0));
-  auto elems  = result.first - keys.ptr(0);
-  auto blocks = get_num_blocks_1d(elems);
-  scatter_reduce<<<blocks, THREADS_PER_BLOCK, 0, stream>>>(elems, out, keys.ptr(0), counts.ptr(0));
-  CHECK_CUDA_STREAM(stream);
-}
-
 __global__ void elementwise_mult_csr_csr_nnz_kernel(size_t rows,
                                                     Rect<1> iterBounds,
                                                     AccessorWO<nnz_ty, 1> nnz,
